Created HomeViewModel and HomeView Refactor
