+++ ProjectSculptExtension/Home/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  ProjectSculptExtension
+//
+//  Created by Monty on 12/3/24.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
