+++ ProjectSculptExtension/Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  ProjectSculptExtension
+//
+//  Created by Monty on 12/3/24.
+//
+
+import SwiftUI
+struct Workout {
+    let id: Int
+    let title: String
+    let image: String
+    let duration: String
+    let date: String
+    let calories: String
+    let tintColor: Color
+    
+}
