+++ ProjectSculptExtension/Home/Views/HomeView.swift
@@ -8,27 +8,8 @@
 import SwiftUI
 
 struct HomeView: View {
-    @State var calories: Int = 123
-    @State var active: Int = 8
-    @State var stand: Int = 4
-    
-    var mockActivities = [
-        Activity( id:0, title:"Todays steps", subtitle:"Goal 12,000", image: "figure.walk", tintColor: .green, amount: "9812"
-                ),
-        Activity( id:1, title:"Todays steps", subtitle:"Goal 11,000", image: "figure.walk", tintColor: .red, amount: "812"
-                ),
-        Activity( id:2, title:"Todays steps", subtitle:"Goal 12,000", image: "figure.walk", tintColor: .blue, amount: "9812"
-                ),
-        Activity( id:3, title:"Todays steps", subtitle:"Goal 12,000", image: "figure.run", tintColor: .purple, amount: "1029"
-                )
-    ]
-    
-    var mockWorkouts = [
-        Workout(id: 0, title: "Chris class", image: "figure.run", duration: "51min", date: "Aug 1", calories: "512 kcal", tintColor: .red),
-        Workout(id: 1, title: "Darrian Class", image: "figure.run", duration: "51min", date: "Aug 11", calories: "512 kcal", tintColor: .purple),
-        Workout(id: 2, title: "Running", image: "figure.run", duration: "51min", date: "Aug 12", calories: "512 kcal", tintColor: .cyan),
-        Workout(id: 3, title: "Running", image: "figure.run", duration: "51min", date: "Aug 13", calories: "512 kcal", tintColor: .cyan),
-    ]
+    @StateObject var viewModel = HomeViewModel()
+   
     
     var body: some View {
         NavigationStack {
@@ -42,14 +23,14 @@ struct HomeView: View {
                         
                         Spacer()
                         
-                        VStack {
+                        VStack(alignment: .leading) {
                             VStack(alignment: .leading, spacing: 8) {
                                 Text("Calories")
                                     .font(.callout)
                                     .bold()
                                     .foregroundColor(.red)
                                 
-                                Text("123 kcal")
+                                Text("\(viewModel.calories)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -60,7 +41,7 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundColor(.red)
                                 
-                                Text("52 mins")
+                                Text("\(viewModel.active)")
                                     .bold()
                             }
                             .padding(.bottom)
@@ -70,19 +51,19 @@ struct HomeView: View {
                                     .bold()
                                     .foregroundColor(.red)
                                 
-                                Text("8 hours")
+                                Text("\(viewModel.stand)")
                                     .bold()
                             }
                         }
                         Spacer()
                         
                         ZStack {
-                            ProgressCircleView(progress: $calories, goal: 600, color: .blue)
+                            ProgressCircleView(progress: $viewModel.calories, goal: 600, color: .blue)
                             
-                            ProgressCircleView(progress: $active, goal: 60, color: .green)
+                            ProgressCircleView(progress: $viewModel.active, goal: 60, color: .green)
                                 .padding(.all, 20)
                             
-                            ProgressCircleView(progress: $stand, goal: 12, color: .blue)
+                            ProgressCircleView(progress: $viewModel.stand, goal: 12, color: .blue)
                                 .padding(.all, 40)
                         }
                         .padding(.horizontal)
@@ -108,7 +89,7 @@ struct HomeView: View {
                     .padding(.horizontal)
                     
                     LazyVGrid(columns: Array(repeating: GridItem(spacing: 20), count: 2)) {
-                        ForEach(mockActivities, id: \.id) { activity in
+                        ForEach(viewModel.mockActivities, id: \.id) { activity in
                             ActivityCard(activity: activity)
                         }
                         
@@ -135,7 +116,7 @@ struct HomeView: View {
                     .padding(.top)
                     
                     LazyVStack {
-                        ForEach(mockWorkouts, id: \.id) { workout in
+                        ForEach(viewModel.mockWorkouts, id: \.id) { workout in
                             WorkoutCard(workout: workout)
                         }
                     }
